WSS: Improve HLS temporary directory handling
- Do not try to cleanup the directory when it has already been deleted.
- Include a random number in the directory name. This fixes access
  problems when seeking in the stream (re-spawning the transcoder).

+++ Services/MPExtended.Services.StreamingService/Units/HTTPLiveStreamUnit.cs
@@ -63,7 +63,8 @@ namespace MPExtended.Services.StreamingService.Units
             {
                 Thread.Sleep(5000);
                 var dir = Path.GetDirectoryName(indexFile);
-                Directory.Delete(dir, true);
+                if (Directory.Exists(dir))
+                    Directory.Delete(dir, true);
             });
 
             return true;
